remove non-conserved test data from tests
this removes all primitive variables from the test output. It removes the odd
timesteps form the shift-atboundary test since the sum(convservatives) is only
constant at even steps when there is no time interpolation for the reduction.
It increases the threshold for the no-refluxing test to value used for other
hydro tests. Ok to apply? With these all tests pass (without being
regenerated).


git-svn-id: https://svn.cct.lsu.edu/repos/numrel/LSUThorns/Refluxing/trunk@74 53d01c76-9b5b-45fe-803d-9063f4e27562